# files/swagger/tenant_schema.yaml
openapi: 3.0.3
info:
  title: Inventory Management System(IMS)
  version: V0.0.0
  description: This Django-based Inventory Management System (IMS) is a scalable,
    multi-tenant backend application built to manage inventory operations across  multiple
    tenants. It supports fine-grained access control, optimized API routing, and production-ready
    deployment with Docker and Nginx.
paths:
...
  /api/user:
    get:
      operationId: user_retrieve
      description: |-
        Retrieve object list based query params provided in the request data.
        Args:
            request (Request): The HTTP request object containing the data.
        Returns:
            object: The list of retrieved objects if found.
      parameters:
      - in: query
        name: first_name
        schema:
          type: string
          minLength: 1
      - in: query
        name: is_pagination
        schema:
          type: boolean
          default: true
      - in: query
        name: last_name
        schema:
          type: string
          minLength: 1
      - in: query
        name: page
        schema:
          type: integer
          minimum: 1
          default: 1
      - in: query
        name: page_size
        schema:
          type: integer
          minimum: 1
          default: 10
      tags:
      - User
      security:
      - TokenAuth: []
      - TokenAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
              examples:
                ListUser-Success:
                  value:
                    data:
                      list:
                      - email: user@example.com
                        user_id: 9d018a56-abd9-4dfd-b606-80ce3ba8f53f
                        profile_photo: https://cdn.example.com/photos/user.jpg
                        phone_number: 9876543210
                        first_name: John
                        last_name: Doe
                        full_name: John Doe
                      pagination:
                        count: 1
                        page_size: 1
                        current_page: 1
                        total_pages: 1
                    errors: null
                    messages: null
                    status_code: 200
                    is_success: true
                  summary: List User - Success
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
              examples:
                404NotFound:
                  value:
                    data: null
                    errors:
                      code: NO_DATA_FOUND
                      message: No Data Found.
                    status_code: 404
                    is_success: false
                  summary: 404 Not Found
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
              examples:
                401Unauthorized:
                  value:
                    data: null
                    errors:
                      code: UNAUTHORIZED
                      message: Unauthorized Access.
                    status_code: 401
                    is_success: false
                  summary: 401 Unauthorized
          description: ''
    post:
      operationId: user_create
      description: |-
        Handles the object creation process, including validation, data enrichment,
        database persistence, and response generation.

        Args:
            request (Request): The HTTP request containing input data.

        Returns:
            Response: A success response with the created object(s)
            or an error message if validation fails.
      tags:
      - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      security:
      - TokenAuth: []
      - TokenAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                CreateUser-Success:
                  value:
                    data:
                      email: user@example.com
                      user_id: 9d018a56-abd9-4dfd-b606-80ce3ba8f53f
                      profile_photo: https://cdn.example.com/photos/user.jpg
                      phone_number: 9876543210
                      first_name: John
                      last_name: Doe
                      full_name: John Doe
                    errors: null
                    messages:
                      message: Created Successfully.
                    status_code: 201
                    is_success: true
                  summary: Create User - Success
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
              examples:
                400BadRequest:
                  value:
                    data: null
                    errors:
                    - code: INVALID
                      message: Invalid data.
                      field: field_name
                    status_code: 400
                    is_success: false
                  summary: 400 Bad Request
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
              examples:
                401Unauthorized:
                  value:
                    data: null
                    errors:
                      code: UNAUTHORIZED
                      message: Unauthorized Access.
                    status_code: 401
                    is_success: false
                  summary: 401 Unauthorized
          description: ''
...
components:
  schemas:
...
    BadRequestResponse:
      type: object
      description: Standard response for a bad request.
      properties:
        data: {}
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
          description: List of validation or business logic errors.
        status_code:
          type: integer
          default: 400
        is_success:
          type: boolean
          default: false
      required:
      - errors
...
    ErrorDetail:
      type: object
      description: Standard error detail format.
      properties:
        code:
          type: string
          description: Machine-readable error code, e.g., 'DUPLICATE_ENTRY'.
        message:
          type: string
          description: Human-readable error message.
        field:
          type: string
          description: Field where the error occurred.
      required:
      - code
      - field
      - message
...
    NotFoundResponse:
      type: object
      description: Standard response for resource not found.
      properties:
        data: {}
        errors:
          allOf:
          - $ref: '#/components/schemas/ErrorDetail'
          description: Resource not found.
        status_code:
          type: integer
          default: 404
        is_success:
          type: boolean
          default: false
      required:
      - errors
...
    Pagination:
      type: object
      description: Pagination for swagger documentation.
      properties:
        count:
          type: integer
        page_size:
          type: integer
        current_page:
          type: integer
        total_pages:
          type: integer
      required:
      - count
      - current_page
      - page_size
      - total_pages
...
    RoleIdEnum:
      enum:
      - SUPER_ADMIN
      - COMPANY_ADMIN
      - MANAGER
      - OPERATOR
      type: string
      description: |-
        * `SUPER_ADMIN` - Super Admin
        * `COMPANY_ADMIN` - Company Admin
        * `MANAGER` - Manager
        * `OPERATOR` - Operator
...
    UnauthorizedResponse:
      type: object
      description: Standard response for unauthorized access.
      properties:
        data: {}
        errors:
          allOf:
          - $ref: '#/components/schemas/ErrorDetail'
          description: Unauthorized
        status_code:
          type: integer
          default: 401
        is_success:
          type: boolean
          default: false
      required:
      - errors
    User:
      type: object
      description: Serializer for both creating and updating a user.
      properties:
        email:
          type: string
          format: email
        phone_number:
          type: integer
        role_id:
          $ref: '#/components/schemas/RoleIdEnum'
        last_name:
          type: string
          maxLength: 16
        first_name:
          type: string
          maxLength: 16
        profile_photo:
          type: string
          format: uri
          nullable: true
        password:
          type: string
          maxLength: 16
          minLength: 4
      required:
      - email
      - first_name
      - last_name
      - password
      - phone_number
      - role_id
    UserData:
      type: object
      description: Serializer for both creating and updating a user.
      properties:
        user_id:
          type: string
          format: uuid
          description: PK for user model
        email:
          type: string
          format: email
          description: User's email address.
        profile_photo:
          type: string
          description: URL or path to the user's profile photo.
        phone_number:
          type: integer
          description: User's phone number.
        last_name:
          type: string
          description: User's last name.
          maxLength: 16
        first_name:
          type: string
          description: User's first name.
          maxLength: 16
      required:
      - email
      - first_name
      - last_name
      - phone_number
      - user_id
    UserListData:
      type: object
      description: Serializer for the data field in user list response.
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/UserData'
          description: List of user records.
        pagination:
          allOf:
          - $ref: '#/components/schemas/Pagination'
          description: Pagination information for the list of users.
      required:
      - list
      - pagination
    UserListResponse:
      type: object
      description: Serializer for the response of the user list endpoint.
      properties:
        data:
          allOf:
          - $ref: '#/components/schemas/UserListData'
          description: Users and pagination.
        errors:
          nullable: true
          description: Any errors message for the response body.
        messages:
          nullable: true
          description: Any informational messages for the response body.
        status_code:
          type: integer
          default: 200
        is_success:
          type: boolean
          default: true
      required:
      - data
      - errors
      - messages
    UserResponse:
      type: object
      description: Serializer for the response of user-related endpoints.
      properties:
        data:
          allOf:
          - $ref: '#/components/schemas/UserData'
          description: ' User Information.'
        errors:
          nullable: true
          description: Any errors message for the response.
        messages:
          nullable: true
          description: Any informational messages for the response body.
        status_code:
          type: integer
          default: 200
        is_success:
          type: boolean
          default: true
      required:
      - data
      - errors
      - messages
  securitySchemes:
    TokenAuth:
      type: http
      scheme: bearer
      bearerFormat: Token